refactor(svm): restructure fee program test suite
