+++ rust/sealevel/programs/hyperlane-sealevel-fee/tests/functional.rs
@@ -1,6 +1,9 @@
 //! Functional tests for the hyperlane-sealevel-fee program.
 
-use std::collections::BTreeSet;
+use std::{
+    collections::BTreeSet,
+    ops::{Deref, DerefMut},
+};
 
 use hyperlane_core::{H160, H256};
 use solana_program::{
@@ -16,13 +19,12 @@ use solana_system_interface::program as system_program;
 
 use serializable_account_meta::SimulationReturnData;
 
-use account_utils::AccountData;
 use hyperlane_sealevel_fee::{
     accounts::{
         CrossCollateralRoute, CrossCollateralRouteAccount, CrossCollateralRoutingFeeConfig,
         FeeAccount, FeeAccountData, FeeData, FeeStandingQuotePda, FeeStandingQuotePdaAccount,
-        FeeStandingQuoteValue, LeafFeeConfig, RouteDomain, RouteDomainAccount, RoutingFeeConfig,
-        DEFAULT_ROUTER, WILDCARD_DOMAIN, WILDCARD_RECIPIENT,
+        LeafFeeConfig, RouteDomain, RouteDomainAccount, RoutingFeeConfig, DEFAULT_ROUTER,
+        WILDCARD_DOMAIN, WILDCARD_RECIPIENT,
     },
     cc_route_pda_seeds,
     error::Error as FeeError,
@@ -45,14 +47,32 @@ fn fee_program_id() -> Pubkey {
 
 // --- Shared test helpers ---
 
-async fn setup_client() -> (BanksClient, Keypair) {
+struct TestBanksClient {
+    ctx: ProgramTestContext,
+}
+
+impl Deref for TestBanksClient {
+    type Target = BanksClient;
+
+    fn deref(&self) -> &Self::Target {
+        &self.ctx.banks_client
+    }
+}
+
+impl DerefMut for TestBanksClient {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.ctx.banks_client
+    }
+}
+
+async fn setup_client() -> (TestBanksClient, Keypair) {
     let program_id = fee_program_id();
     let program_test = ProgramTest::new(
         "hyperlane_sealevel_fee",
         program_id,
         processor!(fee_process_instruction),
     );
-    let mut ctx = program_test.start_with_context().await;
+    let ctx = program_test.start_with_context().await;
     // Reset the clock to 0 so that transient-quote tests using small timestamps work.
     let mut clock = ctx
         .banks_client
@@ -62,10 +82,7 @@ async fn setup_client() -> (BanksClient, Keypair) {
     clock.unix_timestamp = 2;
     ctx.set_sysvar(&clock);
     let payer = ctx.payer.insecure_clone();
-    let banks_client = ctx.banks_client.clone();
-    // Leak the context so the bank task keeps running for the lifetime of the test.
-    Box::leak(Box::new(ctx));
-    (banks_client, payer)
+    (TestBanksClient { ctx }, payer)
 }
 
 async fn process_tx(
@@ -634,16 +651,6 @@ fn build_submit_standing_ix_with_routes(
     .unwrap()
 }
 
-fn build_submit_cc_standing_ix(
-    fee_account: &Pubkey,
-    payer: &Pubkey,
-    quote: &SvmSignedQuote,
-    dest: u32,
-    target_router: &H256,
-) -> Instruction {
-    build_submit_standing_ix_with_routes(fee_account, payer, quote, dest, target_router, &[])
-}
-
 fn build_close_transient_ix(
     fee_key: &Pubkey,
     transient_pda: &Pubkey,
@@ -862,38 +869,6 @@ mod set_beneficiary {
         assert_tx_error(
             result,
             TransactionError::InstructionError(0, InstructionError::InvalidArgument),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::SetBeneficiary(Pubkey::new_unique()),
-            vec![
-                AccountMeta::new(key, false),
-                AccountMeta::new_readonly(payer.pubkey(), true),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
         );
     }
 }
@@ -2319,42 +2294,6 @@ mod add_quote_signer {
             TransactionError::InstructionError(0, InstructionError::InvalidArgument),
         );
     }
-
-    #[tokio::test]
-    async fn test_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::AddQuoteSigner {
-                signer: H160::random(),
-                route: None,
-            },
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
-        );
-    }
 }
 
 mod remove_quote_signer {
@@ -2433,42 +2372,6 @@ mod remove_quote_signer {
         assert_tx_error(
             result,
             TransactionError::InstructionError(0, InstructionError::InvalidArgument),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::RemoveQuoteSigner {
-                signer: H160::random(),
-                route: None,
-            },
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
         );
     }
 }
@@ -2555,38 +2458,6 @@ mod set_min_issued_at {
         assert_tx_error(
             result,
             TransactionError::InstructionError(0, InstructionError::InvalidArgument),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::SetMinIssuedAt { min_issued_at: 100 },
-            vec![
-                AccountMeta::new(fee_key, false),
-                AccountMeta::new_readonly(payer.pubkey(), true),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
         );
     }
 }
@@ -2749,5199 +2620,11 @@ mod set_wildcard_quote_signers {
     }
 }
 
-mod submit_transient_quote {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_submit_transient_quote() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        // Add signer.
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Use a far-future timestamp so the quote hasn't expired.
-        let issued_at = encode_u48(100);
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            vec![1, 2, 3, 4],
-            vec![5, 6, 7, 8],
-            issued_at,
-        );
-
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Verify the transient PDA was created.
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-        let account = banks_client
-            .get_account(transient_pda)
-            .await
-            .unwrap()
-            .unwrap();
-        assert!(!account.data.is_empty());
-        assert_eq!(account.owner, fee_program_id());
-    }
-
-    #[tokio::test]
-    async fn test_invalid_signature_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Create a quote signed by a DIFFERENT key.
-        let wrong_key = SigningKey::random(&mut rand::thread_rng());
-        let issued_at = encode_u48(100);
-        let quote = make_signed_transient_quote(
-            &wrong_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            vec![1, 2, 3],
-            vec![4, 5, 6],
-            issued_at,
-        );
-
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::InvalidQuoteSignature as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_no_signers_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        // No signers added (empty set) — should fail with InvalidQuoteSignature.
-        let issued_at = encode_u48(100);
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            vec![],
-            vec![],
-            issued_at,
-        );
-
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::InvalidQuoteSignature as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_expiry_before_issued_at_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Manually create a quote with expiry < issued_at.
-        let quote = SvmSignedQuote {
-            context: vec![],
-            data: vec![],
-            issued_at: encode_u48(200),
-            expiry: encode_u48(100), // expiry before issued_at
-            client_salt: H256::random(),
-            signature: [0u8; 65],
-        };
-
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::InvalidQuoteExpiry as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let issued_at = encode_u48(100);
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            vec![1, 2],
-            vec![3, 4],
-            issued_at,
-        );
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::SubmitQuote(quote),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new_readonly(fee_key, false),
-                AccountMeta::new(transient_pda, false),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false), // extraneous
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_expired_quote_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Use a very small timestamp that the clock has already passed.
-        let issued_at = encode_u48(1);
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            vec![],
-            vec![],
-            issued_at,
-        );
-
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::QuoteExpired as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_zero_fee_params_transient_quote() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Zero fee params (max_fee=0, half_amount=0).
-        let issued_at = encode_u48(100);
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            vec![0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], // context
-            vec![0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], // data (zero fees)
-            issued_at,
-        );
-
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_double_submit_same_salt_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let issued_at = encode_u48(100);
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            vec![1, 2],
-            vec![3, 4],
-            issued_at,
-        );
-
-        // First submission succeeds.
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Second submission with same quote (same salt → same PDA) should fail.
-        // Use a different payer for the second tx to avoid transaction deduplication.
-        // The quote is still signed for `payer`, so we pass `payer` as extra signer.
-        let payer2 = Keypair::new();
-        fund_keypair(&mut banks_client, &payer, &payer2).await;
-
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        let result = process_tx(&mut banks_client, &payer2, ix, &[&payer]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(0, InstructionError::AccountAlreadyInitialized),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_transient_issued_at_too_far_in_future_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Clock is at 2, MAX_ISSUED_AT_SKEW = 300. issued_at = 400 > 2 + 300 → rejected.
-        let context = encode_context(42, H256::zero(), 100);
-        let data = encode_data(1000, 500);
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(400),
-        );
-
-        let ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::IssuedAtTooFarInFuture as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_standing_issued_at_too_far_in_future_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Standing quote: issued_at = 400 > 2 + 300 → rejected.
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(42, H256::zero()),
-            encode_data(1000, 500),
-            encode_u48(400),        // issued_at: too far in future
-            encode_u48(9999999999), // expiry: far future (valid)
-        );
-
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, 42);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::IssuedAtTooFarInFuture as u32),
-            ),
-        );
-    }
-}
-
-mod quote_fee_transient {
-    use super::*;
-
-    fn build_quote_fee_with_transient_ix(
-        fee_account: &Pubkey,
-        payer: &Pubkey,
-        transient_pda: &Pubkey,
-        dest: u32,
-        recipient: H256,
-        amount: u64,
-    ) -> Instruction {
-        let domain_quotes_pda = standing_quote_pda_for(fee_account, dest);
-        let wildcard_quotes_pda = standing_quote_pda_for(fee_account, WILDCARD_DOMAIN);
-
-        Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::QuoteFee(hyperlane_sealevel_fee::instruction::QuoteFee {
-                destination_domain: dest,
-                recipient,
-                amount,
-                target_router: H256::zero(),
-            }),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new_readonly(*fee_account, false),
-                AccountMeta::new(*payer, true),
-                AccountMeta::new(*transient_pda, false), // transient PDA (writable for autoclose)
-                AccountMeta::new_readonly(domain_quotes_pda, false),
-                AccountMeta::new_readonly(wildcard_quotes_pda, false),
-            ],
-        )
-    }
-
-    #[tokio::test]
-    async fn test_transient_quote_consumed_and_autoclosed() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        // Init fee account with Linear curve, on-chain params max_fee=100, half_amount=50.
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit a transient quote with DIFFERENT params: max_fee=2000, half_amount=1000.
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let amount = 1000u64;
-        let context = encode_context(dest, recipient, amount);
-        let data = encode_data(2000, 1000);
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-
-        let submit_ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        // Derive transient PDA address.
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // Simulate to verify the fee value.
-        let sim_ix = build_quote_fee_with_transient_ix(
-            &fee_key,
-            &payer.pubkey(),
-            &transient_pda,
-            dest,
-            recipient,
-            amount,
-        );
-        let fee = simulate_quote_fee(&mut banks_client, &payer, sim_ix).await;
-
-        // Transient params: Linear max_fee=2000, half_amount=1000, amount=1000.
-        // min(2000, 1000*2000/(2*1000)) = min(2000, 1000) = 1000.
-        // On-chain would give: min(100, 1000*100/(2*50)) = 100. Confirms transient was used.
-        assert_eq!(fee, 1000);
-
-        // Execute to actually close the PDA.
-        let exec_ix = build_quote_fee_with_transient_ix(
-            &fee_key,
-            &payer.pubkey(),
-            &transient_pda,
-            dest,
-            recipient,
-            amount,
-        );
-        process_tx(&mut banks_client, &payer, exec_ix, &[])
-            .await
-            .unwrap();
-
-        // Verify transient PDA was autoclosed.
-        let account = banks_client.get_account(transient_pda).await.unwrap();
-        assert!(
-            account.is_none() || account.unwrap().data.is_empty(),
-            "Transient PDA should be closed after consumption"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_context_mismatch_different_amount() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit transient quote for amount=500.
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let context = encode_context(dest, recipient, 500);
-        let data = encode_data(1000, 500);
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-
-        let submit_ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // QuoteFee with DIFFERENT amount (999 instead of 500).
-        let quote_ix = build_quote_fee_with_transient_ix(
-            &fee_key,
-            &payer.pubkey(),
-            &transient_pda,
-            dest,
-            recipient,
-            999, // mismatch
-        );
-        let result = process_tx(&mut banks_client, &payer, quote_ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::TransientContextMismatch as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_context_mismatch_different_destination() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let context = encode_context(42, H256::zero(), 500);
-        let data = encode_data(1000, 500);
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-
-        let submit_ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // QuoteFee with different destination (99 instead of 42).
-        let quote_ix = build_quote_fee_with_transient_ix(
-            &fee_key,
-            &payer.pubkey(),
-            &transient_pda,
-            99,
-            H256::zero(),
-            500,
-        );
-        let result = process_tx(&mut banks_client, &payer, quote_ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::TransientContextMismatch as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_context_mismatch_different_recipient() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let context = encode_context(42, H256::zero(), 500);
-        let data = encode_data(1000, 500);
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-
-        let submit_ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // QuoteFee with different recipient.
-        let quote_ix = build_quote_fee_with_transient_ix(
-            &fee_key,
-            &payer.pubkey(),
-            &transient_pda,
-            42,
-            H256::random(), // different recipient
-            500,
-        );
-        let result = process_tx(&mut banks_client, &payer, quote_ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::TransientContextMismatch as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_zero_fee_params_consumed() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 1000,
-                    half_amount: 500,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit transient with zero fee params (max_fee=0, half_amount=0).
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let amount = 1000u64;
-        let context = encode_context(dest, recipient, amount);
-        let data = encode_data(0, 0); // zero fee
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-
-        let submit_ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // Simulate to verify the fee value.
-        let sim_ix = build_quote_fee_with_transient_ix(
-            &fee_key,
-            &payer.pubkey(),
-            &transient_pda,
-            dest,
-            recipient,
-            amount,
-        );
-        let fee = simulate_quote_fee(&mut banks_client, &payer, sim_ix).await;
-
-        // Transient params: max_fee=0, half_amount=0 → fee = 0.
-        // On-chain would give: min(1000, 1000*1000/(2*500)) = 1000. Confirms transient was used.
-        assert_eq!(fee, 0);
-
-        // Execute to actually close the PDA.
-        let exec_ix = build_quote_fee_with_transient_ix(
-            &fee_key,
-            &payer.pubkey(),
-            &transient_pda,
-            dest,
-            recipient,
-            amount,
-        );
-        process_tx(&mut banks_client, &payer, exec_ix, &[])
-            .await
-            .unwrap();
-
-        // Verify autoclosed.
-        let account = banks_client.get_account(transient_pda).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_transient_on_routing_account_works() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Routing(RoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        // Configure route for domain 42.
-        let route_strategy = FeeDataStrategy::Regressive(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-        let ix = build_set_route_ix(&fee_key, &payer.pubkey(), 42, route_strategy);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Add signer to route PDA (must exist first).
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::Domain(42)),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let amount = 100u64;
-        let context = encode_context(dest, recipient, amount);
-        let data = encode_data(2000, 1000); // override params
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-
-        let route_pda = route_pda_for(&fee_key, 42);
-        let submit_ix =
-            build_submit_transient_ix_with_routes(&fee_key, &payer.pubkey(), &quote, &[route_pda]);
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // QuoteFee with transient + route PDA. Transient should be consumed.
-        let domain_quotes_pda = standing_quote_pda_for(&fee_key, dest);
-        let wildcard_quotes_pda = standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN);
-        let route_pda = route_pda_for(&fee_key, dest);
-
-        let build_ix = || {
-            Instruction::new_with_borsh(
-                fee_program_id(),
-                &FeeInstruction::QuoteFee(hyperlane_sealevel_fee::instruction::QuoteFee {
-                    destination_domain: dest,
-                    recipient,
-                    amount,
-                    target_router: H256::zero(),
-                }),
-                vec![
-                    AccountMeta::new_readonly(system_program::ID, false),
-                    AccountMeta::new_readonly(fee_key, false),
-                    AccountMeta::new(payer.pubkey(), true),
-                    AccountMeta::new(transient_pda, false),
-                    AccountMeta::new_readonly(domain_quotes_pda, false),
-                    AccountMeta::new_readonly(wildcard_quotes_pda, false),
-                    AccountMeta::new_readonly(route_pda, false),
-                ],
-            )
-        };
-
-        // Simulate to verify the fee value.
-        let fee = simulate_quote_fee(&mut banks_client, &payer, build_ix()).await;
-
-        // Transient params: Regressive max_fee=2000, half_amount=1000, amount=100.
-        // 2000 * 100 / (1000 + 100) = 181.
-        // On-chain route would give: 100 * 100 / (50 + 100) = 66. Confirms transient was used.
-        assert_eq!(fee, 181);
-
-        // Execute to actually close the PDA.
-        process_tx(&mut banks_client, &payer, build_ix(), &[])
-            .await
-            .unwrap();
-
-        // Verify transient PDA was autoclosed.
-        let account = banks_client.get_account(transient_pda).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_transient_on_cc_account_works() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let amount = 500u64;
-        let target_router = H256::random();
-
-        // Configure CC route.
-        let route_strategy = FeeDataStrategy::Progressive(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            target_router,
-            route_strategy,
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Add signer to CC route PDA (must exist first).
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit transient with CC context (76 bytes).
-        let context = encode_cc_context(dest, recipient, amount, target_router);
-        let data = encode_data(3000, 1500);
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-
-        let specific_pda = cc_route_pda_for(&fee_key, dest, &target_router);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let submit_ix = build_submit_transient_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &quote,
-            &[specific_pda, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // QuoteFee with CC accounts + transient.
-        // CC standing PDAs use target_router in seeds.
-        let domain_quotes_pda = cc_standing_quote_pda_for(&fee_key, dest, &target_router);
-        let wildcard_quotes_pda =
-            cc_standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN, &target_router);
-        let cc_specific_pda = cc_route_pda_for(&fee_key, dest, &target_router);
-
-        let build_ix = || {
-            Instruction::new_with_borsh(
-                fee_program_id(),
-                &FeeInstruction::QuoteFee(hyperlane_sealevel_fee::instruction::QuoteFee {
-                    destination_domain: dest,
-                    recipient,
-                    amount,
-                    target_router,
-                }),
-                vec![
-                    AccountMeta::new_readonly(system_program::ID, false),
-                    AccountMeta::new_readonly(fee_key, false),
-                    AccountMeta::new(payer.pubkey(), true),
-                    AccountMeta::new(transient_pda, false),
-                    AccountMeta::new_readonly(domain_quotes_pda, false),
-                    AccountMeta::new_readonly(wildcard_quotes_pda, false),
-                    AccountMeta::new_readonly(cc_specific_pda, false),
-                ],
-            )
-        };
-
-        // Simulate to verify the fee value.
-        let fee = simulate_quote_fee(&mut banks_client, &payer, build_ix()).await;
-
-        // Transient params: Progressive max_fee=3000, half_amount=1500, amount=500.
-        // 3000 * 500^2 / (1500^2 + 500^2) = 300.
-        // On-chain route would give: 100 * 500^2 / (50^2 + 500^2) = 99. Confirms transient was used.
-        assert_eq!(fee, 300);
-
-        // Execute to actually close the PDA.
-        process_tx(&mut banks_client, &payer, build_ix(), &[])
-            .await
-            .unwrap();
-
-        // Verify autoclosed.
-        let account = banks_client.get_account(transient_pda).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_cc_context_wrong_target_router_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let target_router = H256::random();
-        let wrong_router = H256::random();
-
-        let route_strategy = FeeDataStrategy::Linear(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            target_router,
-            route_strategy.clone(),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Also set up a CC route for wrong_router so the transient submit can resolve it.
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            wrong_router,
-            route_strategy,
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Add signer to wrong_router CC route PDA (quote context uses wrong_router).
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: wrong_router,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit transient with wrong target_router in context.
-        let context = encode_cc_context(dest, H256::zero(), 100, wrong_router);
-        let data = encode_data(1000, 500);
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-
-        let wrong_specific_pda = cc_route_pda_for(&fee_key, dest, &wrong_router);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let submit_ix = build_submit_transient_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &quote,
-            &[wrong_specific_pda, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // QuoteFee with correct target_router but quote has wrong_router → mismatch.
-        let domain_quotes_pda = standing_quote_pda_for(&fee_key, dest);
-        let wildcard_quotes_pda = standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN);
-        let cc_specific_pda = cc_route_pda_for(&fee_key, dest, &target_router);
-
-        let quote_ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::QuoteFee(hyperlane_sealevel_fee::instruction::QuoteFee {
-                destination_domain: dest,
-                recipient: H256::zero(),
-                amount: 100,
-                target_router, // correct router in instruction
-            }),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new_readonly(fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(transient_pda, false),
-                AccountMeta::new_readonly(domain_quotes_pda, false),
-                AccountMeta::new_readonly(wildcard_quotes_pda, false),
-                AccountMeta::new_readonly(cc_specific_pda, false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, quote_ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::TransientContextMismatch as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_payer_mismatch_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let amount = 100u64;
-        let context = encode_context(dest, recipient, amount);
-        let data = encode_data(1000, 500);
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(), // signed for payer
-            context,
-            data,
-            issued_at,
-        );
-
-        let submit_ix = build_submit_transient_ix(&fee_key, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        // Different payer tries to consume the transient quote.
-        let other_payer = Keypair::new();
-        fund_keypair(&mut banks_client, &payer, &other_payer).await;
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        let domain_quotes_pda = standing_quote_pda_for(&fee_key, dest);
-        let wildcard_quotes_pda = standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN);
-
-        let quote_ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::QuoteFee(hyperlane_sealevel_fee::instruction::QuoteFee {
-                destination_domain: dest,
-                recipient,
-                amount,
-                target_router: H256::zero(),
-            }),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new_readonly(fee_key, false),
-                AccountMeta::new(other_payer.pubkey(), true), // different payer
-                AccountMeta::new(transient_pda, false),
-                AccountMeta::new_readonly(domain_quotes_pda, false),
-                AccountMeta::new_readonly(wildcard_quotes_pda, false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &other_payer, quote_ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::TransientPayerMismatch as u32),
-            ),
-        );
-    }
-}
-
-mod submit_standing_quote {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_submit_standing_quote() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let context = encode_standing_context(dest, recipient);
-        let data = encode_data(2000, 1000);
-        let issued_at = encode_u48(100);
-        let expiry = encode_u48(9999999999);
-
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-            expiry,
-        );
-
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Verify domain PDA was created with the quote.
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let standing = fetch_standing_pda(&mut banks_client, domain_pda).await;
-        assert_eq!(standing.quotes.len(), 1);
-        let value = standing.quotes.get(&recipient).unwrap();
-        assert_eq!(value.max_fee, 2000);
-        assert_eq!(value.half_amount, 1000);
-        assert_eq!(value.issued_at, 100);
-
-        // Verify domain was added to standing_quote_domains.
-        let fee_acct = fetch_fee_account(&mut banks_client, fee_key).await;
-        assert!(fee_acct.standing_quote_domains.contains(&dest));
-    }
-
-    #[tokio::test]
-    async fn test_replacement_with_newer_issued_at() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-
-        // First quote: issued_at=100, max_fee=1000.
-        let context = encode_standing_context(dest, recipient);
-        let data = encode_data(1000, 500);
-        let quote1 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context.clone(),
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote1, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Second quote: issued_at=200 (newer), max_fee=2000. Should replace.
-        let data2 = encode_data(2000, 1000);
-        let quote2 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data2,
-            encode_u48(200),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote2, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let standing = fetch_standing_pda(&mut banks_client, domain_pda).await;
-        assert_eq!(standing.quotes.len(), 1);
-        let value = standing.quotes.get(&recipient).unwrap();
-        assert_eq!(value.max_fee, 2000);
-        assert_eq!(value.issued_at, 200);
-    }
-
-    #[tokio::test]
-    async fn test_stale_quote_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let context = encode_standing_context(dest, recipient);
-
-        // First quote: issued_at=200.
-        let quote1 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context.clone(),
-            encode_data(1000, 500),
-            encode_u48(200),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote1, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Second quote: issued_at=100 (stale). Should be rejected.
-        let quote2 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            encode_data(2000, 1000),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote2, dest);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::StaleStandingQuote as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_equal_issued_at_is_noop() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let context = encode_standing_context(dest, recipient);
-
-        // First quote: issued_at=100, max_fee=1000.
-        let quote1 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context.clone(),
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote1, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Second quote: same issued_at=100 but max_fee=9999. Should be no-op.
-        let quote2 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            encode_data(9999, 5000),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote2, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Verify original value is unchanged.
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let standing = fetch_standing_pda(&mut banks_client, domain_pda).await;
-        let value = standing.quotes.get(&recipient).unwrap();
-        assert_eq!(value.max_fee, 1000); // not 9999
-    }
-
-    #[tokio::test]
-    async fn test_non_wildcard_amount_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Context with specific amount (not wildcard).
-        let mut context = Vec::with_capacity(44);
-        context.extend_from_slice(&42u32.to_le_bytes());
-        context.extend_from_slice(H256::zero().as_bytes());
-        context.extend_from_slice(&1000u64.to_le_bytes()); // not u64::MAX
-
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, 42);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::StandingQuoteAmountNotWildcard as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_multiple_recipients_per_domain() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let recipient_a = H256::random();
-        let recipient_b = H256::random();
-
-        // Quote for recipient A.
-        let quote_a = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient_a),
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote_a, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Quote for recipient B.
-        let quote_b = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient_b),
-            encode_data(2000, 1000),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote_b, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Verify both entries exist.
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let standing = fetch_standing_pda(&mut banks_client, domain_pda).await;
-        assert_eq!(standing.quotes.len(), 2);
-        assert_eq!(standing.quotes.get(&recipient_a).unwrap().max_fee, 1000);
-        assert_eq!(standing.quotes.get(&recipient_b).unwrap().max_fee, 2000);
-    }
-
-    #[tokio::test]
-    async fn test_wildcard_recipient() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let context =
-            encode_standing_context(dest, hyperlane_sealevel_fee::accounts::WILDCARD_RECIPIENT);
-        let data = encode_data(3000, 1500);
-
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let standing = fetch_standing_pda(&mut banks_client, domain_pda).await;
-        assert!(standing
-            .quotes
-            .contains_key(&hyperlane_sealevel_fee::accounts::WILDCARD_RECIPIENT));
-        assert_eq!(
-            standing
-                .quotes
-                .get(&hyperlane_sealevel_fee::accounts::WILDCARD_RECIPIENT)
-                .unwrap()
-                .max_fee,
-            3000
-        );
-    }
-
-    #[tokio::test]
-    async fn test_zero_fee_params_standing() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let context = encode_standing_context(dest, recipient);
-        let data = encode_data(0, 0); // zero fee
-
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let standing = fetch_standing_pda(&mut banks_client, domain_pda).await;
-        let value = standing.quotes.get(&recipient).unwrap();
-        assert_eq!(value.max_fee, 0);
-        assert_eq!(value.half_amount, 0);
-    }
-
-    #[tokio::test]
-    async fn test_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let dest = 42u32;
-        let context = encode_standing_context(dest, H256::zero());
-        let data = encode_data(1000, 500);
-
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-
-        let domain_le = dest.to_le_bytes();
-        let (domain_pda, _) = Pubkey::find_program_address(
-            fee_standing_quote_pda_seeds!(fee_key, &domain_le),
-            &fee_program_id(),
-        );
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::SubmitQuote(quote),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(fee_key, false),
-                AccountMeta::new(domain_pda, false),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false), // extraneous
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_fully_wildcarded_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let context = encode_standing_context(
-            hyperlane_sealevel_fee::accounts::WILDCARD_DOMAIN,
-            hyperlane_sealevel_fee::accounts::WILDCARD_RECIPIENT,
-        );
-        let data = encode_data(1000, 500);
-
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-
-        let ix = build_submit_standing_ix(
-            &fee_key,
-            &payer.pubkey(),
-            &quote,
-            hyperlane_sealevel_fee::accounts::WILDCARD_DOMAIN,
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::FullyWildcardedStandingQuote as u32),
-            ),
-        );
-    }
-}
-
-mod quote_fee_standing {
-    use super::*;
-
-    async fn setup_with_standing(
-        banks_client: &mut BanksClient,
-        payer: &Keypair,
-        signing_key: &SigningKey,
-        dest: u32,
-        recipient: H256,
-        quoted_max_fee: u64,
-        quoted_half_amount: u64,
-    ) -> Pubkey {
-        let signer_address = eth_address(signing_key);
-        // On-chain Leaf: Linear with max_fee=100, half_amount=50.
-        let fee_key = init_fee_account(
-            banks_client,
-            payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(banks_client, payer, ix, &[]).await.unwrap();
-
-        let quote = make_signed_standing_quote(
-            signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient),
-            encode_data(quoted_max_fee, quoted_half_amount),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, dest);
-        process_tx(banks_client, payer, ix, &[]).await.unwrap();
-
-        fee_key
-    }
-
-    #[tokio::test]
-    async fn test_standing_specific_match_fee_value() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-
-        let dest = 42u32;
-        let recipient = H256::random();
-        // Standing: Linear with max_fee=2000, half_amount=1000.
-        let fee_key = setup_with_standing(
-            &mut banks_client,
-            &payer,
-            &signing_key,
-            dest,
-            recipient,
-            2000,
-            1000,
-        )
-        .await;
-
-        // QuoteFee for amount=1000 → Linear: min(2000, 1000*2000/(2*1000)) = min(2000,1000) = 1000
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, recipient, 1000);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 1000);
-    }
-
-    #[tokio::test]
-    async fn test_no_standing_falls_to_onchain_fee_value() {
-        let (mut banks_client, payer) = setup_client().await;
-
-        // On-chain Leaf: Linear max_fee=100, half_amount=50.
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        // amount=50 → Linear: min(100, 50*100/(2*50)) = min(100, 50) = 50
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), 42, H256::zero(), 50);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 50);
-    }
-
-    #[tokio::test]
-    async fn test_standing_wildcard_recipient_fee_value() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-
-        let dest = 42u32;
-        // Standing with wildcard recipient: max_fee=5000, half_amount=2500.
-        let fee_key = setup_with_standing(
-            &mut banks_client,
-            &payer,
-            &signing_key,
-            dest,
-            WILDCARD_RECIPIENT,
-            5000,
-            2500,
-        )
-        .await;
-
-        // Any recipient should match wildcard. amount=2500 → min(5000, 2500*5000/5000) = 2500
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, H256::random(), 2500);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 2500);
-    }
-
-    #[tokio::test]
-    async fn test_specific_takes_priority_over_wildcard_fee_value() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let dest = 42u32;
-        let specific_recipient = H256::random();
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Wildcard recipient: max_fee=10000.
-        let q1 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, WILDCARD_RECIPIENT),
-            encode_data(10000, 5000),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &q1, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Specific recipient: max_fee=200.
-        let q2 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, specific_recipient),
-            encode_data(200, 100),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &q2, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // amount=100 → specific: min(200, 100*200/200) = 100 (not wildcard: min(10000,...))
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, specific_recipient, 100);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 100);
-    }
-
-    #[tokio::test]
-    async fn test_min_issued_at_skips_stale_falls_to_onchain() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-
-        // Standing: max_fee=9999 (high). On-chain: max_fee=100, half_amount=50.
-        let fee_key = setup_with_standing(
-            &mut banks_client,
-            &payer,
-            &signing_key,
-            dest,
-            recipient,
-            9999,
-            5000,
-        )
-        .await;
-
-        // Bump min_issued_at past the standing quote's issued_at (100).
-        let ix = build_set_min_issued_at_ix(&fee_key, &payer.pubkey(), 200);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // amount=50 → should fall to on-chain: min(100, 50*100/100) = 50
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, recipient, 50);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 50);
-    }
-
-    #[tokio::test]
-    async fn test_standing_zero_fee_value() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-
-        // Standing with zero fee. On-chain has non-zero (100/50).
-        let fee_key = setup_with_standing(
-            &mut banks_client,
-            &payer,
-            &signing_key,
-            dest,
-            recipient,
-            0,
-            0,
-        )
-        .await;
-
-        // Should use standing zero fee, not on-chain 100/50.
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, recipient, 50);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 0);
-    }
-
-    #[tokio::test]
-    async fn test_wildcard_domain_match_fee_value() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let recipient = H256::random();
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(WILDCARD_DOMAIN, recipient),
-            encode_data(8000, 4000),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, WILDCARD_DOMAIN);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // amount=4000 → Linear: min(8000, 4000*8000/8000) = 4000
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), 99, recipient, 4000);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 4000);
-    }
-
-    #[tokio::test]
-    async fn test_domain_standing_takes_priority_over_wildcard_domain() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let dest = 42u32;
-        let recipient = H256::random();
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Domain-specific standing: max_fee=777, half_amount=1 → fee = 777 for amount=100.
-        let q_domain = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient),
-            encode_data(777, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &q_domain, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Wildcard domain standing: max_fee=333, half_amount=1 → fee = 333 for amount=100.
-        let q_wildcard = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(WILDCARD_DOMAIN, recipient),
-            encode_data(333, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &q_wildcard, WILDCARD_DOMAIN);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // amount=100 → domain standing: 777, wildcard standing: 333, on-chain: 100.
-        // Domain-specific must win.
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, recipient, 100);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 777);
-    }
-
-    #[tokio::test]
-    async fn test_transient_takes_priority_over_standing() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let amount = 100u64;
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Standing: max_fee=333, half_amount=1 → fee = min(333, 100*333/2) = 333.
-        let sq = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient),
-            encode_data(333, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &sq, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Transient: max_fee=777, half_amount=1 → fee = min(777, 100*777/2) = 777.
-        let mut transient_ctx = Vec::with_capacity(44);
-        transient_ctx.extend_from_slice(&dest.to_le_bytes());
-        transient_ctx.extend_from_slice(recipient.as_bytes());
-        transient_ctx.extend_from_slice(&amount.to_le_bytes());
-
-        let tq = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            transient_ctx,
-            encode_data(777, 1),
-            encode_u48(100),
-        );
-
-        let scoped_salt = tq.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-        let submit_ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::SubmitQuote(tq),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new_readonly(fee_key, false),
-                AccountMeta::new(transient_pda, false),
-            ],
-        );
-        process_tx(&mut banks_client, &payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        // QuoteFee with transient → transient wins (777/1), not standing (333/1).
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let wildcard_pda = standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN);
-        let quote_ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::QuoteFee(hyperlane_sealevel_fee::instruction::QuoteFee {
-                destination_domain: dest,
-                recipient,
-                amount,
-                target_router: H256::zero(),
-            }),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new_readonly(fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(transient_pda, false),
-                AccountMeta::new_readonly(domain_pda, false),
-                AccountMeta::new_readonly(wildcard_pda, false),
-            ],
-        );
-        let fee = simulate_quote_fee(&mut banks_client, &payer, quote_ix).await;
-        // Transient: Linear max_fee=777, half_amount=1, amount=100 → min(777, 100*777/2) = 777.
-        // Standing would give: min(333, 100*333/2) = 333.
-        // On-chain would give: min(100, 100*100/100) = 100.
-        // All three values are distinct, confirming transient was used.
-        assert_eq!(fee, 777);
-    }
-
-    #[tokio::test]
-    async fn test_spoofed_standing_pda_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-
-        // Create fee account A with on-chain max_fee=100.
-        let salt_a = H256::zero();
-        let fee_key_a = init_fee_account(
-            &mut banks_client,
-            &payer,
-            salt_a,
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key_a, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Create fee account B with a standing quote (max_fee=1, very cheap).
-        let salt_b = H256::random();
-        let fee_key_b = init_fee_account(
-            &mut banks_client,
-            &payer,
-            salt_b,
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 10000,
-                    half_amount: 5000,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key_b, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit cheap standing quote on fee_key_b.
-        let quote_b = make_signed_standing_quote(
-            &signing_key,
-            &fee_key_b,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient),
-            encode_data(1, 1), // very cheap
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key_b, &payer.pubkey(), &quote_b, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Try to use fee_key_b's standing PDA when querying fee_key_a.
-        let spoofed_domain_pda = standing_quote_pda_for(&fee_key_b, dest);
-        let wildcard_pda = standing_quote_pda_for(&fee_key_a, WILDCARD_DOMAIN);
-
-        let quote_ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::QuoteFee(hyperlane_sealevel_fee::instruction::QuoteFee {
-                destination_domain: dest,
-                recipient,
-                amount: 50,
-                target_router: H256::zero(),
-            }),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new_readonly(fee_key_a, false), // querying A
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new_readonly(spoofed_domain_pda, false), // B's PDA!
-                AccountMeta::new_readonly(wildcard_pda, false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, quote_ix, &[]).await;
-        // Should fail because B's PDA doesn't match A's derivation.
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(0, InstructionError::InvalidSeeds),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_expired_standing_falls_to_onchain() {
-        // Standing quote exists but is expired → QuoteFee skips it, uses on-chain params.
-        let program_id = fee_program_id();
-        let program_test = ProgramTest::new(
-            "hyperlane_sealevel_fee",
-            program_id,
-            processor!(fee_process_instruction),
-        );
-        let mut ctx = program_test.start_with_context().await;
-        let payer = ctx.payer.insecure_clone();
-        let banks_client = &mut ctx.banks_client;
-
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let dest = 42u32;
-        let recipient = H256::random();
-
-        // On-chain Leaf: Linear max_fee=100, half_amount=50.
-        let fee_key = init_fee_account(
-            banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Leaf(LeafFeeConfig {
-                strategy: FeeDataStrategy::Linear(FeeParams {
-                    max_fee: 100,
-                    half_amount: 50,
-                }),
-                signers: Some(BTreeSet::new()),
-            }),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Standing quote: max_fee=9999, expiry=5000000000.
-        let sq = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient),
-            encode_data(9999, 5000),
-            encode_u48(100),
-            encode_u48(5000000000),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &sq, dest);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Warp clock past expiry.
-        let mut clock = banks_client
-            .get_sysvar::<solana_program::clock::Clock>()
-            .await
-            .unwrap();
-        clock.unix_timestamp = 6000000000;
-        ctx.set_sysvar(&clock);
-        let banks_client = &mut ctx.banks_client;
-
-        // QuoteFee: standing is expired → falls to on-chain.
-        // On-chain: Linear max_fee=100, half_amount=50, amount=50.
-        // min(100, 50*100/(2*50)) = min(100, 50) = 50.
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, recipient, 50);
-        let fee = simulate_quote_fee(banks_client, &payer, ix).await;
-        assert_eq!(fee, 50);
-    }
-
-    /// Regression: standing quotes remain usable even after the signer set is removed.
-    /// QuoteFee does not check signers — only SubmitQuote does.
-    /// Revocation is via min_issued_at, not signer config changes.
-    #[tokio::test]
-    async fn test_standing_quote_survives_signer_removal() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-
-        // Setup Leaf fee account with standing quote (max_fee=9999, half_amount=5000).
-        let fee_key = setup_with_standing(
-            &mut banks_client,
-            &payer,
-            &signing_key,
-            dest,
-            recipient,
-            9999,
-            5000,
-        )
-        .await;
-
-        // Verify standing quote is used.
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, recipient, 5000);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        // Standing: regressive-like params → fee should come from standing quote, not on-chain.
-        assert!(fee > 0);
-        let fee_before = fee;
-
-        // Remove the signer. FeeAccount.signers becomes Some(empty set).
-        let ix = build_remove_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-        let acct = fetch_fee_account(&mut banks_client, fee_key).await;
-        assert!(leaf_signers(&acct).as_ref().unwrap().is_empty());
-
-        // Standing quote still works — QuoteFee doesn't check signers.
-        let ix = build_quote_fee_leaf_ix(&fee_key, &payer.pubkey(), dest, recipient, 5000);
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, fee_before);
-    }
-}
-
-mod close_transient_quote {
-    use super::*;
-
-    async fn setup_and_submit_transient(
-        banks_client: &mut BanksClient,
-        payer: &Keypair,
-    ) -> (Pubkey, Pubkey) {
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            banks_client,
-            payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(banks_client, payer, ix, &[]).await.unwrap();
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            vec![1, 2, 3, 4],
-            vec![5, 6, 7, 8],
-            encode_u48(100),
-        );
-
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        let submit_ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::SubmitQuote(quote),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new_readonly(fee_key, false),
-                AccountMeta::new(transient_pda, false),
-            ],
-        );
-        process_tx(banks_client, payer, submit_ix, &[])
-            .await
-            .unwrap();
-
-        (fee_key, transient_pda)
-    }
-
-    #[tokio::test]
-    async fn test_close_by_payer() {
-        let (mut banks_client, payer) = setup_client().await;
-        let (fee_key, transient_pda) = setup_and_submit_transient(&mut banks_client, &payer).await;
-
-        let ix = build_close_transient_ix(&fee_key, &transient_pda, &payer.pubkey());
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let account = banks_client.get_account(transient_pda).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_close_by_wrong_payer_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let (fee_key, transient_pda) = setup_and_submit_transient(&mut banks_client, &payer).await;
-
-        let wrong_payer = Keypair::new();
-        fund_keypair(&mut banks_client, &payer, &wrong_payer).await;
-
-        let ix = build_close_transient_ix(&fee_key, &transient_pda, &wrong_payer.pubkey());
-        let result = process_tx(&mut banks_client, &wrong_payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::TransientPayerMismatch as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let (fee_key, transient_pda) = setup_and_submit_transient(&mut banks_client, &payer).await;
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::CloseTransientQuote,
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new_readonly(fee_key, false),
-                AccountMeta::new(transient_pda, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false), // extraneous
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
-        );
-    }
-}
-
-mod prune_expired_quotes {
-    use super::*;
-
-    async fn setup_fee_with_signer(
-        banks_client: &mut BanksClient,
-        payer: &Keypair,
-        signing_key: &SigningKey,
-    ) -> Pubkey {
-        let signer_address = eth_address(signing_key);
-        let fee_key = init_fee_account(
-            banks_client,
-            payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(banks_client, payer, ix, &[]).await.unwrap();
-        fee_key
-    }
-
-    #[tokio::test]
-    async fn test_prune_all_expired_closes_pda() {
-        // Use ProgramTestContext to manipulate the clock.
-        let program_id = fee_program_id();
-        let program_test = ProgramTest::new(
-            "hyperlane_sealevel_fee",
-            program_id,
-            processor!(fee_process_instruction),
-        );
-        let mut ctx = program_test.start_with_context().await;
-        let payer = ctx.payer.insecure_clone();
-        let banks_client = &mut ctx.banks_client;
-
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let fee_key = setup_fee_with_signer(banks_client, &payer, &signing_key).await;
-
-        let dest = 42u32;
-        // Submit with expiry=9999999999 (valid now, will be expired after clock warp).
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, H256::zero()),
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        process_tx(
-            banks_client,
-            &payer,
-            build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, dest),
-            &[],
-        )
-        .await
-        .unwrap();
-
-        // Warp clock past expiry.
-        let mut clock = banks_client
-            .get_sysvar::<solana_program::clock::Clock>()
-            .await
-            .unwrap();
-        clock.unix_timestamp = 99999999999;
-        ctx.set_sysvar(&clock);
-
-        let banks_client = &mut ctx.banks_client;
-
-        let ix = build_prune_ix(&fee_key, &payer.pubkey(), dest);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // PDA should be closed.
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let account = banks_client.get_account(domain_pda).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-
-        // Domain removed from standing_quote_domains.
-        let fee_acct = fetch_fee_account(banks_client, fee_key).await;
-        assert!(!fee_acct.standing_quote_domains.contains(&dest));
-    }
-
-    #[tokio::test]
-    async fn test_prune_keeps_non_expired_entries() {
-        let program_id = fee_program_id();
-        let program_test = ProgramTest::new(
-            "hyperlane_sealevel_fee",
-            program_id,
-            processor!(fee_process_instruction),
-        );
-        let mut ctx = program_test.start_with_context().await;
-        let payer = ctx.payer.insecure_clone();
-        let banks_client = &mut ctx.banks_client;
-
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let fee_key = setup_fee_with_signer(banks_client, &payer, &signing_key).await;
-
-        let dest = 42u32;
-        let expired_recipient = H256::random();
-        let valid_recipient = H256::random();
-
-        // Submit entry that will expire at 50000000000 (before warp target).
-        let q1 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, expired_recipient),
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(50000000000),
-        );
-        process_tx(
-            banks_client,
-            &payer,
-            build_submit_standing_ix(&fee_key, &payer.pubkey(), &q1, dest),
-            &[],
-        )
-        .await
-        .unwrap();
-
-        // Submit entry that won't expire until 200000000000 (after warp target).
-        let q2 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, valid_recipient),
-            encode_data(2000, 1000),
-            encode_u48(100),
-            encode_u48(200000000000),
-        );
-        process_tx(
-            banks_client,
-            &payer,
-            build_submit_standing_ix(&fee_key, &payer.pubkey(), &q2, dest),
-            &[],
-        )
-        .await
-        .unwrap();
-
-        // Warp clock past first entry's expiry but before second's.
-        let mut clock = banks_client
-            .get_sysvar::<solana_program::clock::Clock>()
-            .await
-            .unwrap();
-        clock.unix_timestamp = 99999999999;
-        ctx.set_sysvar(&clock);
-        let banks_client = &mut ctx.banks_client;
-
-        // Prune.
-        let ix = build_prune_ix(&fee_key, &payer.pubkey(), dest);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // PDA should still exist with only the valid entry.
-        let domain_pda = standing_quote_pda_for(&fee_key, dest);
-        let standing = fetch_standing_pda(banks_client, domain_pda).await;
-        assert_eq!(standing.quotes.len(), 1);
-        assert!(!standing.quotes.contains_key(&expired_recipient));
-        assert!(standing.quotes.contains_key(&valid_recipient));
-
-        // Domain still in standing_quote_domains.
-        let fee_acct = fetch_fee_account(banks_client, fee_key).await;
-        assert!(fee_acct.standing_quote_domains.contains(&dest));
-    }
-
-    #[tokio::test]
-    async fn test_prune_nonexistent_domain_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_prune_ix(&fee_key, &payer.pubkey(), 99);
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::RouteNotFound as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_prune_non_owner_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let fee_key = setup_fee_with_signer(&mut banks_client, &payer, &signing_key).await;
-
-        let dest = 42u32;
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, H256::zero()),
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let non_owner = Keypair::new();
-        fund_keypair(&mut banks_client, &payer, &non_owner).await;
-
-        let ix = build_prune_ix(&fee_key, &non_owner.pubkey(), dest);
-        let result = process_tx(&mut banks_client, &non_owner, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(0, InstructionError::InvalidArgument),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_prune_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let fee_key = setup_fee_with_signer(&mut banks_client, &payer, &signing_key).await;
-
-        let dest = 42u32;
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, H256::zero()),
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, dest);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let domain_le = dest.to_le_bytes();
-        let (domain_pda, _) = Pubkey::find_program_address(
-            fee_standing_quote_pda_seeds!(fee_key, &domain_le),
-            &fee_program_id(),
-        );
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::PruneExpiredQuotes {
-                domain: dest,
-                target_router: None,
-            },
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(domain_pda, false),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false), // extraneous
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_prune_none_on_cc_account_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let fee_key = setup_fee_with_signer(&mut banks_client, &payer, &signing_key).await;
-
-        // Re-init as CC. We need a fresh CC fee account.
-        // setup_fee_with_signer creates Leaf. Create a separate CC one.
-        let cc_fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            H256::random(), // different salt
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        // Prune with None (resolves to H256::zero()) on CC account.
-        // No standing PDA was ever created with H256::zero() for CC → RouteNotFound.
-        let domain_le = 42u32.to_le_bytes();
-        let (domain_pda, _) = Pubkey::find_program_address(
-            fee_standing_quote_pda_seeds!(cc_fee_key, &domain_le),
-            &fee_program_id(),
-        );
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::PruneExpiredQuotes {
-                domain: 42,
-                target_router: None,
-            },
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(cc_fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(domain_pda, false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::RouteNotFound as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_prune_some_on_leaf_account_fails() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit a valid Leaf standing quote first so the domain PDA exists.
-        let context = encode_standing_context(42, H256::zero());
-        let data = encode_data(1000, 500);
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &quote, 42);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Prune with Some(random_router) on Leaf account.
-        // PDA derived with random_router was never created → RouteNotFound.
-        let random_router = H256::random();
-        let domain_le = 42u32.to_le_bytes();
-        let (wrong_pda, _) = Pubkey::find_program_address(
-            fee_standing_quote_pda_seeds!(fee_key, &domain_le, random_router),
-            &fee_program_id(),
-        );
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::PruneExpiredQuotes {
-                domain: 42,
-                target_router: Some(random_router),
-            },
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(wrong_pda, false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::RouteNotFound as u32),
-            ),
-        );
-    }
-}
-
-mod get_quote_account_metas {
-    use super::*;
-    use hyperlane_sealevel_fee::accounts::{DEFAULT_ROUTER, WILDCARD_DOMAIN};
-    use serializable_account_meta::SerializableAccountMeta;
-
-    async fn simulate_get_metas(
-        banks_client: &mut BanksClient,
-        payer: &Keypair,
-        fee_account: &Pubkey,
-        destination_domain: u32,
-        target_router: H256,
-        scoped_salt: Option<H256>,
-    ) -> Vec<SerializableAccountMeta> {
-        let instruction = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::GetQuoteAccountMetas(
-                hyperlane_sealevel_fee::instruction::GetQuoteAccountMetas {
-                    destination_domain,
-                    target_router,
-                    scoped_salt,
-                },
-            ),
-            vec![AccountMeta::new_readonly(*fee_account, false)],
-        );
-        let recent_blockhash = banks_client.get_latest_blockhash().await.unwrap();
-        let simulation = banks_client
-            .simulate_transaction(Transaction::new_unsigned(Message::new_with_blockhash(
-                &[instruction],
-                Some(&payer.pubkey()),
-                &recent_blockhash,
-            )))
-            .await
-            .unwrap();
-        if let Some(Err(err)) = &simulation.result {
-            panic!("Simulation failed: {:?}", err);
-        }
-        let return_data = simulation
-            .simulation_details
-            .unwrap()
-            .return_data
-            .expect("no return data");
-        let result: SimulationReturnData<Vec<SerializableAccountMeta>> =
-            borsh::from_slice(&return_data.data).expect("failed to deserialize");
-        result.return_data
-    }
-
-    #[tokio::test]
-    async fn test_leaf_no_transient() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let dest = 42u32;
-        let metas = simulate_get_metas(
-            &mut banks_client,
-            &payer,
-            &fee_key,
-            dest,
-            H256::zero(),
-            None,
-        )
-        .await;
-
-        // prefix (3) + domain_quotes + wildcard_quotes = 5
-        assert_eq!(metas.len(), 5);
-
-        // Fixed prefix.
-        assert_eq!(metas[0].pubkey, system_program::ID);
-        assert_eq!(metas[1].pubkey, fee_key);
-        assert!(metas[2].is_signer);
-        assert!(metas[2].is_writable);
-
-        let expected_domain = standing_quote_pda_for(&fee_key, dest);
-        let expected_wildcard = standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN);
-
-        assert_eq!(metas[3].pubkey, expected_domain);
-        assert!(!metas[3].is_writable);
-
-        assert_eq!(metas[4].pubkey, expected_wildcard);
-        assert!(!metas[4].is_writable);
-    }
-
-    #[tokio::test]
-    async fn test_leaf_with_transient() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let dest = 42u32;
-        let scoped_salt = H256::random();
-
-        let (expected_transient, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        let metas = simulate_get_metas(
-            &mut banks_client,
-            &payer,
-            &fee_key,
-            dest,
-            H256::zero(),
-            Some(scoped_salt),
-        )
-        .await;
-
-        // prefix (3) + transient + domain_quotes + wildcard_quotes = 6
-        assert_eq!(metas.len(), 6);
-
-        assert_eq!(metas[3].pubkey, expected_transient);
-        assert!(metas[3].is_writable);
-        assert!(!metas[3].is_signer);
-
-        assert_eq!(metas[4].pubkey, standing_quote_pda_for(&fee_key, dest));
-        assert_eq!(
-            metas[5].pubkey,
-            standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_routing() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Routing(RoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let metas = simulate_get_metas(
-            &mut banks_client,
-            &payer,
-            &fee_key,
-            dest,
-            H256::zero(),
-            None,
-        )
-        .await;
-
-        // prefix (3) + domain_quotes + wildcard_quotes + route_pda = 6
-        assert_eq!(metas.len(), 6);
-
-        assert_eq!(metas[3].pubkey, standing_quote_pda_for(&fee_key, dest));
-        assert_eq!(
-            metas[4].pubkey,
-            standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN)
-        );
-        assert_eq!(metas[5].pubkey, route_pda_for(&fee_key, dest));
-        assert!(!metas[5].is_writable);
-    }
-
-    #[tokio::test]
-    async fn test_cc_routing() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let target_router = H256::random();
-        let metas = simulate_get_metas(
-            &mut banks_client,
-            &payer,
-            &fee_key,
-            dest,
-            target_router,
-            None,
-        )
-        .await;
-
-        // prefix (3) + domain_quotes + wildcard_quotes + cc_specific + cc_default = 7
-        assert_eq!(metas.len(), 7);
-
-        assert_eq!(
-            metas[3].pubkey,
-            cc_standing_quote_pda_for(&fee_key, dest, &target_router)
-        );
-        assert_eq!(
-            metas[4].pubkey,
-            cc_standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN, &target_router)
-        );
-        assert_eq!(
-            metas[5].pubkey,
-            cc_route_pda_for(&fee_key, dest, &target_router)
-        );
-        assert_eq!(
-            metas[6].pubkey,
-            cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_cc_routing_with_transient() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let target_router = H256::random();
-        let scoped_salt = H256::random();
-
-        let (expected_transient, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-
-        let metas = simulate_get_metas(
-            &mut banks_client,
-            &payer,
-            &fee_key,
-            dest,
-            target_router,
-            Some(scoped_salt),
-        )
-        .await;
-
-        // prefix (3) + transient + domain_quotes + wildcard_quotes + cc_specific + cc_default = 8
-        assert_eq!(metas.len(), 8);
-
-        assert_eq!(metas[3].pubkey, expected_transient);
-        assert!(metas[3].is_writable);
-        assert_eq!(
-            metas[4].pubkey,
-            cc_standing_quote_pda_for(&fee_key, dest, &target_router)
-        );
-        assert_eq!(
-            metas[5].pubkey,
-            cc_standing_quote_pda_for(&fee_key, WILDCARD_DOMAIN, &target_router)
-        );
-        assert_eq!(
-            metas[6].pubkey,
-            cc_route_pda_for(&fee_key, dest, &target_router)
-        );
-        assert_eq!(
-            metas[7].pubkey,
-            cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_extraneous_account_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::GetQuoteAccountMetas(
-                hyperlane_sealevel_fee::instruction::GetQuoteAccountMetas {
-                    destination_domain: 42,
-                    target_router: H256::zero(),
-                    scoped_salt: None,
-                },
-            ),
-            vec![
-                AccountMeta::new_readonly(fee_key, false),
-                AccountMeta::new_readonly(Pubkey::new_unique(), false), // extraneous
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ExtraneousAccount as u32),
-            ),
-        );
-    }
-}
-
-mod get_submit_quote_account_metas {
-    use super::*;
-    use hyperlane_sealevel_fee::accounts::DEFAULT_ROUTER;
-    use serializable_account_meta::SerializableAccountMeta;
-
-    async fn simulate_get_submit_metas(
-        banks_client: &mut BanksClient,
-        payer: &Keypair,
-        fee_account: &Pubkey,
-        destination_domain: u32,
-        target_router: H256,
-        scoped_salt: Option<H256>,
-    ) -> Vec<SerializableAccountMeta> {
-        let instruction = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::GetSubmitQuoteAccountMetas(
-                hyperlane_sealevel_fee::instruction::GetSubmitQuoteAccountMetas {
-                    destination_domain,
-                    target_router,
-                    scoped_salt,
-                },
-            ),
-            vec![AccountMeta::new_readonly(*fee_account, false)],
-        );
-        let recent_blockhash = banks_client.get_latest_blockhash().await.unwrap();
-        let simulation = banks_client
-            .simulate_transaction(Transaction::new_unsigned(Message::new_with_blockhash(
-                &[instruction],
-                Some(&payer.pubkey()),
-                &recent_blockhash,
-            )))
-            .await
-            .unwrap();
-        if let Some(Err(err)) = &simulation.result {
-            panic!("Simulation failed: {:?}", err);
-        }
-        let return_data = simulation
-            .simulation_details
-            .unwrap()
-            .return_data
-            .expect("no return data");
-        let result: SimulationReturnData<Vec<SerializableAccountMeta>> =
-            borsh::from_slice(&return_data.data).expect("failed to deserialize");
-        result.return_data
-    }
-
-    #[tokio::test]
-    async fn test_leaf_transient() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let scoped_salt = H256::random();
-        let metas = simulate_get_submit_metas(
-            &mut banks_client,
-            &payer,
-            &fee_key,
-            42,
-            H256::zero(),
-            Some(scoped_salt),
-        )
-        .await;
-
-        // system + payer + fee_account + transient_pda = 4
-        assert_eq!(metas.len(), 4);
-        assert_eq!(metas[0].pubkey, system_program::ID);
-        assert!(metas[1].is_signer); // payer
-        assert!(!metas[2].is_writable); // fee_account read-only for transient
-        assert!(metas[3].is_writable); // transient PDA
-    }
-
-    #[tokio::test]
-    async fn test_leaf_standing() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let metas =
-            simulate_get_submit_metas(&mut banks_client, &payer, &fee_key, 42, H256::zero(), None)
-                .await;
-
-        // system + payer + fee_account(W) + standing_pda = 4
-        assert_eq!(metas.len(), 4);
-        assert!(metas[2].is_writable); // fee_account writable for standing
-        assert!(metas[3].is_writable); // standing PDA
-    }
-
-    #[tokio::test]
-    async fn test_routing_transient() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::Routing(RoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let ix = build_set_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            FeeDataStrategy::Linear(FeeParams {
-                max_fee: 1000,
-                half_amount: 500,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let scoped_salt = H256::random();
-        let metas = simulate_get_submit_metas(
-            &mut banks_client,
-            &payer,
-            &fee_key,
-            dest,
-            H256::zero(),
-            Some(scoped_salt),
-        )
-        .await;
-
-        // system + payer + fee_account + route_domain_pda + transient_pda = 5
-        assert_eq!(metas.len(), 5);
-        assert_eq!(metas[3].pubkey, route_pda_for(&fee_key, dest));
-        assert!(!metas[3].is_writable); // route PDA read-only
-        assert!(metas[4].is_writable); // transient PDA
-    }
-
-    #[tokio::test]
-    async fn test_cc_standing() {
-        let (mut banks_client, payer) = setup_client().await;
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let target_router = H256::random();
-
-        let metas = simulate_get_submit_metas(
-            &mut banks_client,
-            &payer,
-            &fee_key,
-            dest,
-            target_router,
-            None,
-        )
-        .await;
-
-        // system + payer + fee_account(W) + cc_specific + cc_default + standing_pda = 6
-        assert_eq!(metas.len(), 6);
-        assert_eq!(
-            metas[3].pubkey,
-            cc_route_pda_for(&fee_key, dest, &target_router)
-        );
-        assert_eq!(
-            metas[4].pubkey,
-            cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER)
-        );
-        assert!(metas[5].is_writable); // standing PDA
-    }
-}
-
-mod get_program_version {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_returns_version() {
-        let (mut banks_client, payer) = setup_client().await;
-
-        let instruction = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::GetProgramVersion,
-            vec![],
-        );
-        let recent_blockhash = banks_client.get_latest_blockhash().await.unwrap();
-        let simulation = banks_client
-            .simulate_transaction(Transaction::new_unsigned(Message::new_with_blockhash(
-                &[instruction],
-                Some(&payer.pubkey()),
-                &recent_blockhash,
-            )))
-            .await
-            .unwrap();
-
-        assert!(simulation.result.unwrap().is_ok());
-        let return_data = simulation
-            .simulation_details
-            .unwrap()
-            .return_data
-            .expect("no return data");
-        let result: SimulationReturnData<String> =
-            borsh::from_slice(&return_data.data).expect("failed to deserialize");
-        assert_eq!(result.return_data, package_versioned::PACKAGE_VERSION);
-    }
-}
-
-mod cc_standing_quotes {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_cc_standing_creates_route_bound_pda() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let target_router = H256::random();
-        let recipient = H256::zero();
-
-        // Set up CC route so the route PDA exists for signer lookup.
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            target_router,
-            FeeDataStrategy::Linear(FeeParams {
-                max_fee: 1000,
-                half_amount: 500,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Add signer to CC route PDA (must exist first).
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let context = encode_cc_standing_context(dest, recipient, target_router);
-        let data = encode_data(2000, 1000);
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-
-        let specific_pda = cc_route_pda_for(&fee_key, dest, &target_router);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &quote,
-            dest,
-            &target_router,
-            &[specific_pda, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Verify PDA is at the CC-specific address.
-        let cc_pda = cc_standing_quote_pda_for(&fee_key, dest, &target_router);
-        let standing = fetch_standing_pda(&mut banks_client, cc_pda).await;
-        assert_eq!(standing.quotes.len(), 1);
-        assert_eq!(standing.quotes.get(&recipient).unwrap().max_fee, 2000);
-    }
-
-    #[tokio::test]
-    async fn test_two_routers_get_separate_pdas() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let router_a = H256::random();
-        let router_b = H256::random();
-
-        // Set up CC routes for both routers.
-        let route_strategy = FeeDataStrategy::Linear(FeeParams {
-            max_fee: 1000,
-            half_amount: 500,
-        });
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            router_a,
-            route_strategy.clone(),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, router_b, route_strategy);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Add signer to both CC route PDAs.
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: router_a,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: router_b,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit standing for router_a: max_fee=1000.
-        let ctx_a = encode_cc_standing_context(dest, recipient, router_a);
-        let q_a = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            ctx_a,
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda_a = cc_route_pda_for(&fee_key, dest, &router_a);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &q_a,
-            dest,
-            &router_a,
-            &[specific_pda_a, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit standing for router_b: max_fee=5000.
-        let ctx_b = encode_cc_standing_context(dest, recipient, router_b);
-        let q_b = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            ctx_b,
-            encode_data(5000, 2500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda_b = cc_route_pda_for(&fee_key, dest, &router_b);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &q_b,
-            dest,
-            &router_b,
-            &[specific_pda_b, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Verify separate PDAs with separate values.
-        let pda_a = cc_standing_quote_pda_for(&fee_key, dest, &router_a);
-        let pda_b = cc_standing_quote_pda_for(&fee_key, dest, &router_b);
-        assert_ne!(pda_a, pda_b);
-
-        let standing_a = fetch_standing_pda(&mut banks_client, pda_a).await;
-        let standing_b = fetch_standing_pda(&mut banks_client, pda_b).await;
-        assert_eq!(standing_a.quotes.get(&recipient).unwrap().max_fee, 1000);
-        assert_eq!(standing_b.quotes.get(&recipient).unwrap().max_fee, 5000);
-    }
-
-    #[tokio::test]
-    async fn test_cc_standing_zero_target_router_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        // Set up a DEFAULT_ROUTER CC route so signer resolution succeeds via fallback.
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            42,
-            DEFAULT_ROUTER,
-            FeeDataStrategy::Linear(FeeParams {
-                max_fee: 1000,
-                half_amount: 500,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: 42,
-                target_router: DEFAULT_ROUTER,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit with target_router = H256::zero() → should be rejected.
-        let context = encode_cc_standing_context(42, H256::zero(), H256::zero());
-        let data = encode_data(1000, 500);
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-
-        let specific_pda = cc_route_pda_for(&fee_key, 42, &H256::zero());
-        let default_pda = cc_route_pda_for(&fee_key, 42, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &quote,
-            42,
-            &H256::zero(),
-            &[specific_pda, default_pda],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ZeroTargetRouterNotAllowed as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_cc_standing_default_router_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        // Set up a DEFAULT_ROUTER CC route so signer resolution succeeds.
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            42,
-            DEFAULT_ROUTER,
-            FeeDataStrategy::Linear(FeeParams {
-                max_fee: 1000,
-                half_amount: 500,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: 42,
-                target_router: DEFAULT_ROUTER,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit with target_router = DEFAULT_ROUTER → should be rejected.
-        let context = encode_cc_standing_context(
-            42,
-            H256::zero(),
-            hyperlane_sealevel_fee::accounts::DEFAULT_ROUTER,
-        );
-        let data = encode_data(1000, 500);
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-
-        let specific_pda = cc_route_pda_for(&fee_key, 42, &DEFAULT_ROUTER);
-        let default_pda = cc_route_pda_for(&fee_key, 42, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &quote,
-            42,
-            &hyperlane_sealevel_fee::accounts::DEFAULT_ROUTER,
-            &[specific_pda, default_pda],
-        );
-        let result = process_tx(&mut banks_client, &payer, ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::ZeroTargetRouterNotAllowed as u32),
-            ),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_cc_prune_does_not_remove_domain_from_tracking() {
-        let program_id = fee_program_id();
-        let program_test = ProgramTest::new(
-            "hyperlane_sealevel_fee",
-            program_id,
-            processor!(fee_process_instruction),
-        );
-        let mut ctx = program_test.start_with_context().await;
-        let payer = ctx.payer.insecure_clone();
-        let banks_client = &mut ctx.banks_client;
-
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        // Set up a CC route so QuoteFee can resolve the curve.
-        let dest = 42u32;
-        let router_a = H256::random();
-        let router_b = H256::random();
-        let route_strategy = FeeDataStrategy::Linear(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            router_a,
-            route_strategy.clone(),
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, router_b, route_strategy);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Add signer to both CC route PDAs.
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: router_a,
-            }),
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: router_b,
-            }),
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Submit standing quotes for two different routers on the same domain.
-        let ctx_a = encode_cc_standing_context(dest, H256::zero(), router_a);
-        let q_a = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            ctx_a,
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda_a = cc_route_pda_for(&fee_key, dest, &router_a);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &q_a,
-            dest,
-            &router_a,
-            &[specific_pda_a, default_pda],
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        let ctx_b = encode_cc_standing_context(dest, H256::zero(), router_b);
-        let q_b = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            ctx_b,
-            encode_data(2000, 1000),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda_b = cc_route_pda_for(&fee_key, dest, &router_b);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &q_b,
-            dest,
-            &router_b,
-            &[specific_pda_b, default_pda],
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // CC accounts do not track domains in standing_quote_domains.
-        let fee_acct = fetch_fee_account(banks_client, fee_key).await;
-        assert!(!fee_acct.standing_quote_domains.contains(&dest));
-
-        // Warp clock past expiry — both quotes have expiry=9999999999.
-        let mut clock = banks_client
-            .get_sysvar::<solana_program::clock::Clock>()
-            .await
-            .unwrap();
-        clock.unix_timestamp = 99999999999;
-        ctx.set_sysvar(&clock);
-        let banks_client = &mut ctx.banks_client;
-
-        // Prune router_a's standing PDA — it should close.
-        let domain_le = dest.to_le_bytes();
-        let (pda_a, _) = Pubkey::find_program_address(
-            fee_standing_quote_pda_seeds!(fee_key, &domain_le, router_a),
-            &fee_program_id(),
-        );
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::PruneExpiredQuotes {
-                domain: dest,
-                target_router: Some(router_a),
-            },
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(pda_a, false),
-            ],
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Router_a's PDA should be closed.
-        let account = banks_client.get_account(pda_a).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-
-        // CC accounts never add domains to standing_quote_domains,
-        // so the set remains empty regardless of prune operations.
-        let fee_acct = fetch_fee_account(banks_client, fee_key).await;
-        assert!(!fee_acct.standing_quote_domains.contains(&dest));
-    }
-
-    #[tokio::test]
-    async fn test_cc_standing_quote_consumed_in_quote_fee() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let recipient = H256::random();
-        let target_router = H256::random();
-
-        // Configure CC route: Progressive max_fee=100, half_amount=50.
-        let route_strategy = FeeDataStrategy::Progressive(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            target_router,
-            route_strategy,
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Add signer to CC route PDA (must exist first).
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit CC standing quote: max_fee=888, half_amount=1.
-        let context = encode_cc_standing_context(dest, recipient, target_router);
-        let data = encode_data(888, 1);
-        let quote = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda = cc_route_pda_for(&fee_key, dest, &target_router);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &quote,
-            dest,
-            &target_router,
-            &[specific_pda, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // QuoteFee with CC standing quote.
-        let amount = 200u64;
-        let ix = build_quote_fee_cc_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            recipient,
-            amount,
-            target_router,
-            false,
-        );
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-
-        // Standing params: Progressive max_fee=888, half_amount=1, amount=200.
-        // 888 * 200^2 / (1^2 + 200^2) = 888 * 40000 / 40001 = 887.
-        // On-chain route would give: 100 * 200^2 / (50^2 + 200^2) = 100*40000/42500 = 94.
-        // Confirms standing was used.
-        assert_eq!(fee, 887);
-    }
-}
-
-mod additional_coverage {
-    use super::*;
-
-    // --- Test 1: CC standing quotes are router-bound in QuoteFee ---
-
-    #[tokio::test]
-    async fn test_cc_quote_fee_uses_router_bound_standing_pda() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let recipient = H256::random();
-        let router_a = H256::random();
-        let router_b = H256::random();
-
-        // CC routes for both routers — same curve type, same on-chain params.
-        let route = FeeDataStrategy::Linear(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, router_a, route.clone());
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, router_b, route);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Add signer to both CC route PDAs.
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: router_a,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: router_b,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Standing for router_a: max_fee=777, half_amount=1.
-        let sq_a = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_cc_standing_context(dest, recipient, router_a),
-            encode_data(777, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda_a = cc_route_pda_for(&fee_key, dest, &router_a);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &sq_a,
-            dest,
-            &router_a,
-            &[specific_pda_a, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Standing for router_b: max_fee=333, half_amount=1.
-        let sq_b = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_cc_standing_context(dest, recipient, router_b),
-            encode_data(333, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda_b = cc_route_pda_for(&fee_key, dest, &router_b);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &sq_b,
-            dest,
-            &router_b,
-            &[specific_pda_b, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let amount = 100u64;
-
-        // QuoteFee for router_a → should use 777/1: min(777, 100*777/2) = 777.
-        let ix = build_quote_fee_cc_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            recipient,
-            amount,
-            router_a,
-            false,
-        );
-        let fee_a = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee_a, 777);
-
-        // QuoteFee for router_b → should use 333/1: min(333, 100*333/2) = 333.
-        let ix = build_quote_fee_cc_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            recipient,
-            amount,
-            router_b,
-            false,
-        );
-        let fee_b = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee_b, 333);
-    }
-
-    // --- Test 2: CC wildcard-domain standing quote is router-scoped ---
-
-    #[tokio::test]
-    async fn test_cc_quote_fee_wildcard_domain_fallback_is_router_scoped() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        // Init with wildcard signers so wildcard-domain quotes can be submitted.
-        let mut wildcard_signers = BTreeSet::new();
-        wildcard_signers.insert(signer_address);
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: wildcard_signers,
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let recipient = H256::random();
-        let router_a = H256::random();
-        let router_b = H256::random();
-
-        // CC routes for both routers: on-chain Linear max_fee=100, half_amount=50.
-        let route = FeeDataStrategy::Linear(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, router_a, route.clone());
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, router_b, route);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Wildcard-domain standing for router_a ONLY: max_fee=555, half_amount=1.
-        // Auth comes from fee_data.wildcard_signers (no route PDAs needed).
-        let sq = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_cc_standing_context(WILDCARD_DOMAIN, recipient, router_a),
-            encode_data(555, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &sq,
-            WILDCARD_DOMAIN,
-            &router_a,
-            &[], // No route PDAs — wildcard auth from fee_data
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let amount = 100u64;
-
-        // QuoteFee router_a → wildcard standing: min(555, 100*555/2) = 555.
-        let ix = build_quote_fee_cc_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            recipient,
-            amount,
-            router_a,
-            false,
-        );
-        let fee_a = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee_a, 555);
-
-        // QuoteFee router_b → no standing quote exists, falls to on-chain: min(100, 100*100/100) = 100.
-        let ix = build_quote_fee_cc_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            recipient,
-            amount,
-            router_b,
-            false,
-        );
-        let fee_b = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee_b, 100);
-    }
-
-    // --- Test 3: CC exact recipient beats wildcard recipient ---
-
-    #[tokio::test]
-    async fn test_cc_quote_fee_exact_recipient_beats_wildcard_recipient() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let target_router = H256::random();
-        let exact_recipient = H256::random();
-        let other_recipient = H256::random();
-
-        let route = FeeDataStrategy::Linear(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, target_router, route);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Add signer to CC route PDA (must exist first).
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router,
-            }),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit exact-recipient standing: max_fee=888, half_amount=1.
-        let sq_exact = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_cc_standing_context(dest, exact_recipient, target_router),
-            encode_data(888, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda = cc_route_pda_for(&fee_key, dest, &target_router);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &sq_exact,
-            dest,
-            &target_router,
-            &[specific_pda, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Submit wildcard-recipient standing (same PDA, different key): max_fee=444, half_amount=1.
-        let sq_wildcard = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_cc_standing_context(dest, WILDCARD_RECIPIENT, target_router),
-            encode_data(444, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &sq_wildcard,
-            dest,
-            &target_router,
-            &[specific_pda, default_pda],
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let amount = 100u64;
-
-        // QuoteFee for exact_recipient → exact match: min(888, 100*888/2) = 888.
-        let ix = build_quote_fee_cc_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            exact_recipient,
-            amount,
-            target_router,
-            false,
-        );
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 888);
-
-        // QuoteFee for other_recipient → wildcard match: min(444, 100*444/2) = 444.
-        let ix = build_quote_fee_cc_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            other_recipient,
-            amount,
-            target_router,
-            false,
-        );
-        let fee = simulate_quote_fee(&mut banks_client, &payer, ix).await;
-        assert_eq!(fee, 444);
-    }
-
-    // --- Test 4: CC prune one router preserves other router's quote ---
-
-    #[tokio::test]
-    async fn test_cc_prune_one_router_preserves_other_router_quote() {
-        let program_id = fee_program_id();
-        let program_test = ProgramTest::new(
-            "hyperlane_sealevel_fee",
-            program_id,
-            processor!(fee_process_instruction),
-        );
-        let mut ctx = program_test.start_with_context().await;
-        let payer = ctx.payer.insecure_clone();
-        let banks_client = &mut ctx.banks_client;
-
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let recipient = H256::random();
-        let router_a = H256::random();
-        let router_b = H256::random();
-
-        let route = FeeDataStrategy::Linear(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, router_a, route.clone());
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, router_b, route);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Add signer to both CC route PDAs.
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: router_a,
-            }),
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-        let ix = build_add_quote_signer_ix_with_route(
-            &fee_key,
-            &payer.pubkey(),
-            signer_address,
-            Some(instruction::RouteKey::CrossCollateral {
-                destination: dest,
-                target_router: router_b,
-            }),
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Router_a standing: expires at 5000000000 (will be expired).
-        let sq_a = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_cc_standing_context(dest, recipient, router_a),
-            encode_data(777, 1),
-            encode_u48(100),
-            encode_u48(5000000000),
-        );
-        let specific_pda_a = cc_route_pda_for(&fee_key, dest, &router_a);
-        let default_pda = cc_route_pda_for(&fee_key, dest, &DEFAULT_ROUTER);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &sq_a,
-            dest,
-            &router_a,
-            &[specific_pda_a, default_pda],
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Router_b standing: expires at 9999999999 (will be live).
-        let sq_b = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_cc_standing_context(dest, recipient, router_b),
-            encode_data(333, 1),
-            encode_u48(100),
-            encode_u48(9999999999),
-        );
-        let specific_pda_b = cc_route_pda_for(&fee_key, dest, &router_b);
-        let ix = build_submit_standing_ix_with_routes(
-            &fee_key,
-            &payer.pubkey(),
-            &sq_b,
-            dest,
-            &router_b,
-            &[specific_pda_b, default_pda],
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Warp clock past router_a's expiry but before router_b's.
-        let mut clock = banks_client
-            .get_sysvar::<solana_program::clock::Clock>()
-            .await
-            .unwrap();
-        clock.unix_timestamp = 6000000000;
-        ctx.set_sysvar(&clock);
-        let banks_client = &mut ctx.banks_client;
-
-        // Prune router_a.
-        let pda_a = cc_standing_quote_pda_for(&fee_key, dest, &router_a);
-        let ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::PruneExpiredQuotes {
-                domain: dest,
-                target_router: Some(router_a),
-            },
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new(fee_key, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(pda_a, false),
-            ],
-        );
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Router_a PDA closed.
-        let account = banks_client.get_account(pda_a).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-
-        // Router_b PDA still exists.
-        let pda_b = cc_standing_quote_pda_for(&fee_key, dest, &router_b);
-        let standing_b = fetch_standing_pda(banks_client, pda_b).await;
-        assert_eq!(standing_b.quotes.len(), 1);
-        assert_eq!(standing_b.quotes.get(&recipient).unwrap().max_fee, 333);
-
-        // QuoteFee for router_b still returns standing fee: min(333, 100*333/2) = 333.
-        let ix = build_quote_fee_cc_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            recipient,
-            100,
-            router_b,
-            false,
-        );
-        let fee = simulate_quote_fee(banks_client, &payer, ix).await;
-        assert_eq!(fee, 333);
-
-        // CC accounts never track domains in standing_quote_domains.
-        let fee_acct = fetch_fee_account(banks_client, fee_key).await;
-        assert!(!fee_acct.standing_quote_domains.contains(&dest));
-    }
-
-    // --- Test 5: Transient PDA spoof rejected across fee accounts ---
-
-    #[tokio::test]
-    async fn test_transient_pda_spoof_rejected() {
-        let (mut banks_client, payer) = setup_client().await;
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let salt_a = H256::zero();
-        let salt_b = H256::repeat_byte(0x01);
-
-        // Fee account A.
-        let fee_key_a = init_fee_account(
-            &mut banks_client,
-            &payer,
-            salt_a,
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-        let ix = build_add_quote_signer_ix(&fee_key_a, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Fee account B.
-        let fee_key_b = init_fee_account(
-            &mut banks_client,
-            &payer,
-            salt_b,
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-        let ix = build_add_quote_signer_ix(&fee_key_b, &payer.pubkey(), signer_address);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Create transient quote under fee account B.
-        let dest = 42u32;
-        let recipient = H256::zero();
-        let amount = 100u64;
-        let context = encode_context(dest, recipient, amount);
-        let data = encode_data(5000, 1);
-        let issued_at = encode_u48(100);
-
-        let quote = make_signed_transient_quote(
-            &signing_key,
-            &fee_key_b,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            context,
-            data,
-            issued_at,
-        );
-        let ix = build_submit_transient_ix(&fee_key_b, &payer.pubkey(), &quote);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        // Derive B's transient PDA.
-        let scoped_salt = quote.compute_scoped_salt(&payer.pubkey());
-        let (transient_pda_b, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key_b, scoped_salt),
-            &fee_program_id(),
-        );
-
-        // Try QuoteFee on fee account A with B's transient PDA → TransientPdaMismatch.
-        let domain_quotes_pda = standing_quote_pda_for(&fee_key_a, dest);
-        let wildcard_quotes_pda = standing_quote_pda_for(&fee_key_a, WILDCARD_DOMAIN);
-        let quote_ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::QuoteFee(hyperlane_sealevel_fee::instruction::QuoteFee {
-                destination_domain: dest,
-                recipient,
-                amount,
-                target_router: H256::zero(),
-            }),
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new_readonly(fee_key_a, false),
-                AccountMeta::new(payer.pubkey(), true),
-                AccountMeta::new(transient_pda_b, false),
-                AccountMeta::new_readonly(domain_quotes_pda, false),
-                AccountMeta::new_readonly(wildcard_quotes_pda, false),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, quote_ix, &[]).await;
-        assert_tx_error(
-            result,
-            TransactionError::InstructionError(
-                0,
-                InstructionError::Custom(FeeError::TransientPdaMismatch as u32),
-            ),
-        );
-
-        // Try CloseTransientQuote on fee account A with B's transient PDA → should fail.
-        let close_ix = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::CloseTransientQuote,
-            vec![
-                AccountMeta::new_readonly(system_program::ID, false),
-                AccountMeta::new_readonly(fee_key_a, false),
-                AccountMeta::new(transient_pda_b, false),
-                AccountMeta::new(payer.pubkey(), true),
-            ],
-        );
-        let result = process_tx(&mut banks_client, &payer, close_ix, &[]).await;
-        // CloseTransientQuote re-derives PDA from fee_account + scoped_salt.
-        // Since fee_key_a != fee_key_b, derivation won't match.
-        assert!(result.is_err());
-    }
-
-    // --- Test 6: Standing PDA recreated after prune ---
-
-    #[tokio::test]
-    async fn test_standing_pda_recreated_after_prune() {
-        let program_id = fee_program_id();
-        let program_test = ProgramTest::new(
-            "hyperlane_sealevel_fee",
-            program_id,
-            processor!(fee_process_instruction),
-        );
-        let mut ctx = program_test.start_with_context().await;
-        let payer = ctx.payer.insecure_clone();
-        let banks_client = &mut ctx.banks_client;
-
-        let signing_key = SigningKey::random(&mut rand::thread_rng());
-        let signer_address = eth_address(&signing_key);
-
-        let fee_key = init_fee_account(
-            banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-        let ix = build_add_quote_signer_ix(&fee_key, &payer.pubkey(), signer_address);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        let dest = 42u32;
-        let recipient = H256::random();
-
-        // Submit standing quote with expiry=5000000000.
-        let sq1 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient),
-            encode_data(1000, 500),
-            encode_u48(100),
-            encode_u48(5000000000),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &sq1, dest);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // Verify PDA exists and domain tracked.
-        let pda_key = standing_quote_pda_for(&fee_key, dest);
-        let standing = fetch_standing_pda(banks_client, pda_key).await;
-        assert_eq!(standing.quotes.len(), 1);
-        let fee_acct = fetch_fee_account(banks_client, fee_key).await;
-        assert!(fee_acct.standing_quote_domains.contains(&dest));
-
-        // Warp clock past expiry and prune.
-        let mut clock = banks_client
-            .get_sysvar::<solana_program::clock::Clock>()
-            .await
-            .unwrap();
-        clock.unix_timestamp = 6000000000;
-        ctx.set_sysvar(&clock);
-        let banks_client = &mut ctx.banks_client;
-
-        let ix = build_prune_ix(&fee_key, &payer.pubkey(), dest);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // PDA closed, domain removed.
-        let account = banks_client.get_account(pda_key).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-        let fee_acct = fetch_fee_account(banks_client, fee_key).await;
-        assert!(!fee_acct.standing_quote_domains.contains(&dest));
-
-        // Submit a new standing quote on the same domain — PDA should be recreated.
-        let sq2 = make_signed_standing_quote(
-            &signing_key,
-            &fee_key,
-            LOCAL_DOMAIN,
-            &payer.pubkey(),
-            encode_standing_context(dest, recipient),
-            encode_data(2000, 1000),
-            encode_u48(5500000000),
-            encode_u48(9999999999),
-        );
-        let ix = build_submit_standing_ix(&fee_key, &payer.pubkey(), &sq2, dest);
-        process_tx(banks_client, &payer, ix, &[]).await.unwrap();
-
-        // PDA recreated with new params.
-        let standing = fetch_standing_pda(banks_client, pda_key).await;
-        assert_eq!(standing.quotes.len(), 1);
-        assert_eq!(standing.quotes.get(&recipient).unwrap().max_fee, 2000);
-
-        // Domain re-added to tracking.
-        let fee_acct = fetch_fee_account(banks_client, fee_key).await;
-        assert!(fee_acct.standing_quote_domains.contains(&dest));
-    }
-
-    // --- Test 7: CC route recreated after remove ---
-
-    #[tokio::test]
-    async fn test_cc_route_recreated_after_remove() {
-        let (mut banks_client, payer) = setup_client().await;
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            FeeData::CrossCollateralRouting(CrossCollateralRoutingFeeConfig {
-                wildcard_signers: BTreeSet::new(),
-            }),
-        )
-        .await;
-
-        let dest = 42u32;
-        let target_router = H256::random();
-        let strategy1 = FeeDataStrategy::Linear(FeeParams {
-            max_fee: 100,
-            half_amount: 50,
-        });
-
-        // Create route.
-        let ix = build_set_cc_route_ix(&fee_key, &payer.pubkey(), dest, target_router, strategy1);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let pda = cc_route_pda_for(&fee_key, dest, &target_router);
-        let route = fetch_cc_route(&mut banks_client, pda).await;
-        assert_eq!(route.fee_data.params().max_fee, 100);
-
-        // Remove route.
-        let ix = build_remove_cc_route_ix(&fee_key, &payer.pubkey(), dest, target_router);
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let account = banks_client.get_account(pda).await.unwrap();
-        assert!(account.is_none() || account.unwrap().data.is_empty());
-
-        // Recreate with different params.
-        let strategy2 = FeeDataStrategy::Regressive(FeeParams {
-            max_fee: 999,
-            half_amount: 333,
-        });
-        let ix = build_set_cc_route_ix(
-            &fee_key,
-            &payer.pubkey(),
-            dest,
-            target_router,
-            strategy2.clone(),
-        );
-        process_tx(&mut banks_client, &payer, ix, &[])
-            .await
-            .unwrap();
-
-        let route = fetch_cc_route(&mut banks_client, pda).await;
-        assert_eq!(route.fee_data, strategy2);
-    }
-
-    // --- Test 8: GetQuoteAccountMetas scoped_salt is same-tx only ---
-
-    /// GetQuoteAccountMetas with scoped_salt=Some(...) returns a transient PDA account meta.
-    /// This is only valid when SubmitQuote runs in the same tx before QuoteFee.
-    /// If used standalone (without a preceding SubmitQuote), the transient PDA will be
-    /// uninitialized and QuoteFee will skip it — falling through to standing/on-chain.
-    /// This test documents that intended behavior.
-    #[tokio::test]
-    async fn test_get_quote_account_metas_scoped_salt_same_tx_expectation() {
-        use serializable_account_meta::SerializableAccountMeta;
-
-        let (mut banks_client, payer) = setup_client().await;
-
-        let fee_key = init_fee_account(
-            &mut banks_client,
-            &payer,
-            default_salt(),
-            Some(payer.pubkey()),
-            payer.pubkey(),
-            default_leaf_fee_data(),
-        )
-        .await;
-
-        let scoped_salt = H256::random();
-
-        let instruction = Instruction::new_with_borsh(
-            fee_program_id(),
-            &FeeInstruction::GetQuoteAccountMetas(
-                hyperlane_sealevel_fee::instruction::GetQuoteAccountMetas {
-                    destination_domain: 42,
-                    target_router: H256::zero(),
-                    scoped_salt: Some(scoped_salt),
-                },
-            ),
-            vec![AccountMeta::new_readonly(fee_key, false)],
-        );
-
-        let recent_blockhash = banks_client.get_latest_blockhash().await.unwrap();
-        let simulation = banks_client
-            .simulate_transaction(Transaction::new_unsigned(Message::new_with_blockhash(
-                &[instruction],
-                Some(&payer.pubkey()),
-                &recent_blockhash,
-            )))
-            .await
-            .unwrap();
-        assert!(simulation.result.as_ref().unwrap().is_ok());
-
-        let return_data = simulation
-            .simulation_details
-            .unwrap()
-            .return_data
-            .expect("no return data");
-        let result: SimulationReturnData<Vec<SerializableAccountMeta>> =
-            borsh::from_slice(&return_data.data).expect("failed to deserialize");
-        let metas = result.return_data;
-
-        // With scoped_salt=Some, transient PDA is included (writable) at index 3.
-        // Callers MUST pair this with a SubmitQuote in the same tx — the transient PDA
-        // is created by SubmitQuote and consumed by QuoteFee within a single transaction.
-        let (expected_transient, _) = Pubkey::find_program_address(
-            transient_quote_pda_seeds!(fee_key, scoped_salt),
-            &fee_program_id(),
-        );
-        assert_eq!(metas[3].pubkey, expected_transient);
-        assert!(metas[3].is_writable);
-
-        // Without SubmitQuote preceding QuoteFee, the PDA is uninitialized and QuoteFee
-        // will skip it, falling through to standing quote or on-chain resolution.
-    }
-}
+#[path = "functional/cc.rs"]
+mod cc;
+#[path = "functional/helpers.rs"]
+mod helpers;
+#[path = "functional/standing.rs"]
+mod standing;
+#[path = "functional/transient.rs"]
+mod transient;
